+++ pint2vela/tests/test_data_files.py
@@ -18,6 +18,7 @@ datasets = [
     "sim4",
     "sim_fd",
     # "sim6",
+    "J0613-0200.InPTA.NB",
 ]