+++ src/model/measurement_noise.jl
@@ -25,7 +25,7 @@ correct_toa(wn::MeasurementNoise, ctoa::CorrectedTOA, params::NamedTuple) =
     correct_toa(ctoa; efac = efac(wn, ctoa, params), equad2 = equad2(wn, ctoa, params))
 
 function show(io::IO, wn::MeasurementNoise)
-    num_efacs = length(unique(wn.efac_index_mask))
-    num_equads = length(unique(wn.equad_index_mask))
+    num_efacs = length(filter(x -> x > 0, unique(wn.efac_index_mask)))
+    num_equads = length(filter(x -> x > 0, unique(wn.equad_index_mask)))
     print(io, "MeasurementNoise($num_efacs EFACs, $num_equads EQUADs)")
 end