+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Variable-index chromatic delay as a Taylor expansion (`ChromaticTaylor`)
 - Auto-generation of HTML documentation using `Documenter`
 - `docs-CI` tests
-- Examples - `sim1`, `sim_jump`, `sim_jump_ex`, `sim_fdjumpdm`, `sim_sw`, `sim_cm`, `sim_fd`
+- Examples - `sim1`, `sim_jump`, `sim_jump_ex`, `sim_fdjumpdm`, `sim_sw`, `sim_cm`, `sim_fd`, `J0613-0200.InPTA.NB`
 - Tests corresponding to the example datasets
 - System-dependent DM offsets (`DispersionOffset`)
 - `compare_residuals.py` script in `examples`.
@@ -23,7 +23,9 @@
 - Fourier series representation of achromatic red noise (`WaveX`), DM noise (`DMWaveX`), and chromatic noise (`ChromaticCM`)
 - `get_lnpost_func` function
 - Frequency-dependent profile variability corrections (`FrequencyDependent`)
-- Memory allocation tests for all componentss
+- Memory allocation tests for all components
+- `mean_anomaly` and `mean_motion` functions
+- `BinaryELL1` model
 ## Changed
 - Reorganized source files into subdirectories
 - Replaced `par` and `tim` files for testing with `JLSO` files
@@ -32,6 +34,7 @@
 - Merged `pint2vela` into the main repo.
 - Replaced `pint2vela` test datasets with symlinks.
 ## Fixed
+- `show` method for `MeasurementNoise`
 ## Removed
 - `read_model_and_toas()` function. Data is now read from `JLSO` files created using `pint2vela`
 - `plot_summary()` function. (This is better done in Python.)